V3.5.1 #230 #261 #269

+++ SBFspot/SBFspot.h
@@ -448,7 +448,7 @@ const char *getEventGroup(unsigned long eGroup);
 const char *getEventType(unsigned short eventflags);
 int getInverterData(InverterData *inverters[], enum getInverterDataType type);
 int getInverterIndexByAddress(InverterData *inverters[], unsigned char bt_addr[6]);
-int getInverterIndexBySerial(InverterData *inverters[], unsigned short SUSyID, unsigned long Serial);
+int getInverterIndexBySerial(InverterData *inverters[], unsigned short SUSyID, uint32_t Serial);
 int getInverterIndexBySerial(InverterData *inverters[], unsigned long Serial);
 E_SBFSPOT getPacket(unsigned char senderaddr[6], int wait4Command);
 //int get_tzOffset(void);
